energia-iot-backend/requirements.txt
# Dependencias - Plataforma IoT Medidor de Energía
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
paho-mqtt>=2.0.0
psycopg2-binary>=2.9.0
